Fix problem with content_length not being a long which caused problems with long files.  Fixed decoding of URL on daemon so that it unescapes characters.  Fixed repository problems crashing because of a null lock object.  Cleanup empty temp directories on startup.  Fix exceptions with sync source or target dirs cannot be resolved.  Changed disconnect log warning to a trace statement because it is normal.

+++ HttpServer/HttpRequest.cs
@@ -384,8 +384,12 @@ namespace HttpServer
                         AcceptTypes[i] = AcceptTypes[i].Trim();
                     break;
                 case "content-length":
-                    int t;
-                    if (!int.TryParse(value, out t))
+                    // DMO
+                    //int t;
+                    //if (!int.TryParse(value, out t))
+                    //    throw new BadRequestException("Invalid content length.");
+                    long t;
+                    if (!long.TryParse(value, out t))
                         throw new BadRequestException("Invalid content length.");
                     ContentLength = t;
                     break; //todo: mayby throw an exception
